refactored components
refactored the following files: Modals, DeployedScripts, FilesDeleteAll, FilesUpload, InfoModal, KissVMFileCreate, KissVMFileDelete, KissVMFileDeleteAll, KissVMFileRename, ModalBase, ModalNotFound, WorkspaceCopy, WorkspaceCreateBlank, WorkspaceDelete, WorkspaceDeleteAll, WorkspaceExport, WorkspaceImport, WorkspaceRename, LeftSidePanel, BasicHighlight, BasicInput, defaults, and components. added new components: BasicHeadings, BasicProgressBar, and DragDropInput.

+++ src/components/ui/modals/FilesDeleteAll.tsx
@@ -1,35 +1,40 @@
-import { Box, Highlight, Input, Text } from '@chakra-ui/react';
-import ConfirmModal from './ConfirmModal';
+// Import dependencies
+import { Box, Text } from '@chakra-ui/react';
 import { useState } from 'react';
+// Import stores
 import useFileStore from '@/stores/useFileStore';
-import useAppTheme from '@/themes/useAppTheme';
+import useModalStore from '@/stores/useModalStore';
+// Import constants
+import { CONFIRM_TEXTS, INPUT_PLACEHOLDERS } from '@/constants';
+// Import components
+import BasicHighlight from '../systems/BasicHighlight';
+import BasicInput from '../systems/BasicInput';
+import ConfirmModal from './ConfirmModal';
 
-// Constants
-const CONFIRM_TEXT = 'Delete all files';
-
-// Workspace rename modal component
-function FilesDeleteAll({ onClose }) {
+// Delete all files modal component
+function FilesDeleteAll() {
   // Define store
   const deleteAllFiles = useFileStore((state) => state.deleteAllFiles);
+  const onClose = useModalStore((state) => state.onClose);
 
   // Define state
   const [confirmText, setConfirmText] = useState('');
 
-  // Define theme
-  const { accent, color, colorAlt, borderColor } = useAppTheme();
+  // Define handlers
+  function handleOnClick() {
+    deleteAllFiles();
+    setConfirmText('');
+    onClose();
+  }
 
   // Render
   return (
     <ConfirmModal
-      title="Delete all workspaces"
+      title="Delete all files"
       buttonLabel="Confirm"
       onClose={onClose}
-      onClick={() => {
-        deleteAllFiles();
-        setConfirmText('');
-        onClose();
-      }}
-      disabled={confirmText !== CONFIRM_TEXT.toUpperCase()}
+      onClick={handleOnClick}
+      disabled={confirmText !== CONFIRM_TEXTS.deleteAllFiles}
     >
       <Text fontSize="sm" pb={4} textAlign="center">
         Are you sure you want to delete all files? This action cannot be undone.
@@ -37,36 +42,15 @@ function FilesDeleteAll({ onClose }) {
         <span className="uppercase">uppercase</span> to confirm.
       </Text>
 
-      <Text pb={4} textAlign="center">
-        <Highlight
-          query={CONFIRM_TEXT.toUpperCase()}
-          styles={{
-            px: '3',
-            py: '1',
-            rounded: 'sm',
-            color: 'gray.50',
-            bg: 'red.700',
-            userSelect: 'none',
-          }}
-        >
-          {CONFIRM_TEXT.toUpperCase()}
-        </Highlight>
-      </Text>
+      <BasicHighlight pb={4} query={CONFIRM_TEXTS.deleteAllFiles || '---'} />
 
       <Box px={4}>
-        <Input
-          size="sm"
-          variant="outline"
-          borderColor={borderColor}
-          _placeholder={{ color: borderColor }}
-          _hover={{ borderColor: color }}
-          _focusVisible={{
-            borderColor: accent,
-          }}
-          _readOnly={{ color: colorAlt }}
+        <BasicInput
+          placeholder={INPUT_PLACEHOLDERS.confirm}
           value={confirmText}
-          onChange={(e) => setConfirmText(e.target.value)}
-          placeholder="Enter confirm text here"
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+            setConfirmText(e.target.value)
+          }
         />
       </Box>
     </ConfirmModal>
