+++ src/components/ui/modals/WorkspaceDeleteAll.tsx
@@ -1,23 +1,36 @@
+// Import dependencies
 import { Box, Text } from '@chakra-ui/react';
-import ConfirmModal from './ConfirmModal';
 import { useState } from 'react';
+// Import stores
+import useModalStore from '@/stores/useModalStore';
 import useWorkspaceStore from '@/stores/useWorkspaceStore';
-import BasicInput from '../systems/BasicInput';
+// Import constants
+import { CONFIRM_TEXTS, INPUT_PLACEHOLDERS } from '@/constants';
+// Import components
+import ConfirmModal from './ConfirmModal';
 import BasicHighlight from '../systems/BasicHighlight';
-import { INPUT_PLACEHOLDERS } from '@/constants';
+import BasicInput from '../systems/BasicInput';
 
-// Constants
-const CONFIRM_TEXT = 'Delete all workspaces';
-
-// Workspace rename modal component
-function WorkspaceDeleteAll({ onClose }) {
+// Delete all workspaces modal component
+function WorkspaceDeleteAll() {
   // Define store
   const deleteAllWorkspaces = useWorkspaceStore(
     (state) => state.deleteAllWorkspaces
   );
+  const onClose = useModalStore((state) => state.onClose);
 
   // Define state
   const [confirmText, setConfirmText] = useState('');
+
+  // Define handlers
+  function handleOnClick() {
+    deleteAllWorkspaces();
+    setConfirmText('');
+    onClose();
+  }
+  function handleOnChange(e: React.ChangeEvent<HTMLInputElement>) {
+    setConfirmText(e.target.value);
+  }
 
   // Render
   return (
@@ -25,26 +38,25 @@ function WorkspaceDeleteAll({ onClose }) {
       title="Delete all workspaces"
       buttonLabel="Confirm"
       onClose={onClose}
-      onClick={() => {
-        deleteAllWorkspaces();
-        setConfirmText('');
-        onClose();
-      }}
-      disabled={confirmText !== CONFIRM_TEXT.toUpperCase()}
+      onClick={handleOnClick}
+      disabled={confirmText !== CONFIRM_TEXTS.deleteAllWorkspaces}
     >
       <Text fontSize="sm" pb={4} textAlign="center">
         Are you sure you want to delete all workspaces? This action cannot be
-        undone. Please type the confirm text in{' '}
+        undone. Please type the confirm text in&nbsp;
         <span className="uppercase">uppercase</span> to confirm.
       </Text>
 
-      <BasicHighlight pb={4} query={CONFIRM_TEXT.toUpperCase()} />
+      <BasicHighlight
+        pb={4}
+        query={CONFIRM_TEXTS.deleteAllWorkspaces || '---'}
+      />
 
       <Box px={4}>
         <BasicInput
           placeholder={INPUT_PLACEHOLDERS.confirm}
           value={confirmText}
-          onChange={(e) => setConfirmText(e.target.value)}
+          onChange={handleOnChange}
         />
       </Box>
     </ConfirmModal>
